LUCENE-8958: Concurrent Early Termination for TopScoreDocCollector
